move proposal and vote package into types.
