+++ consensus/propose.go
@@ -1,8 +1,8 @@
 package consensus
 
 import (
-	"github.com/zarbchain/zarb-go/consensus/proposal"
-	"github.com/zarbchain/zarb-go/consensus/vote"
+	"github.com/zarbchain/zarb-go/types/proposal"
+	"github.com/zarbchain/zarb-go/types/vote"
 )
 
 type proposeState struct {
