+++ consensus/voteset/block_votes.go
@@ -1,7 +1,7 @@
 package voteset
 
 import (
-	"github.com/zarbchain/zarb-go/consensus/vote"
+	"github.com/zarbchain/zarb-go/types/vote"
 	"github.com/zarbchain/zarb-go/types/crypto"
 	"github.com/zarbchain/zarb-go/util/logger"
 )
