+++ sync/bundle/message/proposal_test.go
@@ -4,7 +4,7 @@ import (
 	"testing"
 
 	"github.com/stretchr/testify/assert"
-	"github.com/zarbchain/zarb-go/consensus/proposal"
+	"github.com/zarbchain/zarb-go/types/proposal"
 	"github.com/zarbchain/zarb-go/util/errors"
 )
