+++ types/proposal/proposal.go
